Revision numbers are expected to be integers

+++ src/main/java/org/radarcns/management/repository/CustomRevisionEntityRepository.java
@@ -3,5 +3,6 @@ package org.radarcns.management.repository;
 import org.radarcns.management.domain.audit.CustomRevisionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomRevisionEntityRepository extends JpaRepository<CustomRevisionEntity, Long> {
+public interface CustomRevisionEntityRepository extends JpaRepository<CustomRevisionEntity,
+        Integer> {
 }
